BigInteger

+++ FibonacciSequence.Tests/FibonacciGeneratorTests.cs
@@ -13,7 +13,7 @@ namespace FibonacciSequence.Tests
         {
             FibonacciGenerator generator = new FibonacciGenerator();
 
-            List<double> sequence = generator.Generate1(10);
+            List<BigInteger> sequence = generator.Generate1(10);
 
             Assert.AreEqual(0, sequence[0]);
             Assert.AreEqual(1, sequence[1]);
@@ -29,7 +29,7 @@ namespace FibonacciSequence.Tests
         {
             FibonacciGenerator generator = new FibonacciGenerator();
 
-            List<double> sequence = generator.Generate2(10);
+            List<BigInteger> sequence = generator.Generate2(10);
 
             Assert.AreEqual(0, sequence[0]);
             Assert.AreEqual(1, sequence[1]);
@@ -45,7 +45,7 @@ namespace FibonacciSequence.Tests
         {
             FibonacciGenerator generator = new FibonacciGenerator();
 
-            List<double> sequence = generator.Generate3(10);
+            List<BigInteger> sequence = generator.Generate3(10);
 
             Assert.AreEqual(0, sequence[0]);
             Assert.AreEqual(1, sequence[1]);
@@ -60,7 +60,7 @@ namespace FibonacciSequence.Tests
         {
             FibonacciGenerator generator = new FibonacciGenerator();
 
-            List<double> sequence = generator.Generate4(10);
+            List<BigInteger> sequence = generator.Generate4(10);
 
             Assert.AreEqual(0, sequence[0]);
             Assert.AreEqual(1, sequence[1]);
@@ -75,7 +75,7 @@ namespace FibonacciSequence.Tests
         {
             FibonacciGenerator generator = new FibonacciGenerator();
 
-            List<double> sequence = generator.Generate5(47);
+            List<BigInteger> sequence = generator.Generate5(47);
 
             Assert.AreEqual(0, sequence[0]);
             Assert.AreEqual(1, sequence[1]);
@@ -91,7 +91,7 @@ namespace FibonacciSequence.Tests
         {
             FibonacciGenerator generator = new FibonacciGenerator();
 
-            List<double> sequence = generator.Generate6YieldToList();
+            List<BigInteger> sequence = generator.Generate6YieldToList();
 
             Assert.AreEqual(0, sequence[0]);
             Assert.AreEqual(1, sequence[1]);
@@ -107,7 +107,7 @@ namespace FibonacciSequence.Tests
         {
             FibonacciGenerator generator = new FibonacciGenerator();
 
-            List<double> sequence = generator.Generat7YieldTake();
+            List<BigInteger> sequence = generator.Generat7YieldTake();
 
             Assert.AreEqual(0, sequence[0]);
             Assert.AreEqual(1, sequence[1]);
@@ -125,13 +125,13 @@ namespace FibonacciSequence.Tests
 
             ArrayList sequence = generator.Generate8();
 
-            Assert.AreEqual((double)0, sequence[0]);
-            Assert.AreEqual((double)1, sequence[1]);
-            Assert.AreEqual((double)1, sequence[2]);
-            Assert.AreEqual((double)2, sequence[3]);
-            Assert.AreEqual((double)3, sequence[4]);
-            Assert.AreEqual((double)5, sequence[5]);
-            Assert.AreEqual((double)8, sequence[6]);
+            Assert.AreEqual((BigInteger)0, sequence[0]);
+            Assert.AreEqual((BigInteger)1, sequence[1]);
+            Assert.AreEqual((BigInteger)1, sequence[2]);
+            Assert.AreEqual((BigInteger)2, sequence[3]);
+            Assert.AreEqual((BigInteger)3, sequence[4]);
+            Assert.AreEqual((BigInteger)5, sequence[5]);
+            Assert.AreEqual((BigInteger)8, sequence[6]);
         }
 
         [TestMethod]
@@ -139,7 +139,7 @@ namespace FibonacciSequence.Tests
         {
             FibonacciGenerator generator = new FibonacciGenerator();
 
-            double[] sequence = generator.Generate9();
+            BigInteger[] sequence = generator.Generate9();
 
             Assert.AreEqual(0, sequence[0]);
             Assert.AreEqual(1, sequence[1]);
@@ -155,7 +155,7 @@ namespace FibonacciSequence.Tests
         {
             FibonacciGenerator generator = new FibonacciGenerator();
 
-            Span<double> sequence = generator.Generate10();
+            Span<BigInteger> sequence = generator.Generate10();
 
             Assert.AreEqual(0, sequence[0]);
             Assert.AreEqual(1, sequence[1]);
@@ -172,7 +172,7 @@ namespace FibonacciSequence.Tests
         {
             FibonacciGenerator generator = new FibonacciGenerator();
 
-            double[] sequence = generator.Generate11();
+            BigInteger[] sequence = generator.Generate11();
 
             Assert.AreEqual(0, sequence[0]);
             Assert.AreEqual(1, sequence[1]);
@@ -188,7 +188,7 @@ namespace FibonacciSequence.Tests
         {
             FibonacciGenerator generator = new FibonacciGenerator();
 
-            List<double> sequence = generator.Generate12();
+            List<BigInteger> sequence = generator.Generate12();
 
             Assert.AreEqual(0, sequence[0]);
             Assert.AreEqual(1, sequence[1]);
